requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
billiard==3.6.3.0
celery==4.4.7
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
defusedxml==0.6.0
distlib==0.3.1
Django==2.2.16
django-allauth==0.42.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.11.1
filelock==3.0.12
flake8==3.8.3
idna==2.10
importlib-metadata==2.1.1
importlib-resources==3.2.1
isort==4.3.21
kombu==4.6.11
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
oauthlib==3.1.0
pep8==1.7.1
Pillow==5.4.1
pipenv==2020.11.15
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.6.0
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.2
typed-ast==1.4.1
urllib3==1.25.11
vine==1.3.0
virtualenv==20.2.1
virtualenv-clone==0.5.4
wrapt==1.12.1
zipp==1.2.0
